+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+### Version 0.3.0
+### Breaking
+- Remove `Into`/`From` implementations for `Color`
+- Remove specific conversions method (such as `_as_f32_array`)
+- Add `to_rgb`, `from_rgb`, `to_argb`, `from_argb`, `to_rgba`, `from_rgba` for `u32`, `(u8,u8,u8,u8)`, `[u8;4]`, `(u8,u8,u8)`, `[u8;3]`, `(f32,f32,f32,f32)`, `[f32;4]`, `(f32,f32,f32)`, `[f32;3]`
+  - This was as it was confusing which format the conversion methods worked with
+- Migration:
+  - For `let num: u32 = color.into()` use `let num: u32 = color.to_rgba()` (or `to_argb()` as needed)
+  - For `let arr = color.as_f32_array()` use `let arr: [f32; 4] = color.to_rgba()`
+  - For `let color: Color = 1_u32.into()` use `let color = Color::from_rgba(1_u32)`
+
 ### Version 0.2.5
 - Update docs
 - Update deps